[#77] extract one of the loader functions
